Separate controller and middleware from route.
